# api/api.yaml
swagger: '2.0'
...
info:
  version: "1.0.0"
  title: ASW Issue Tracker API
  description: |
    The first version of the ASW API is an exciting step forward towards
    making it easier for users to have open access to their data. We created it
    so that you can surface the amazing content ASW-Project users share.

    Build something great!

    All endpoints are only accessible via http and are located at
    `api.aswproject.fun`. 
    
    You must use an API Key for each endpoint, the new user creation *POST /users* doesn't require it though.
    
    In order to get yours, you should LogIn here [api.aswproject.fun/login](http://api.aswproject.fun/login), once you get authenticated with your email and password, a message will appear telling you your API Key, copy it.
    To use your API Key, click on the green Authenticate button below, on the Security section and paste it there and click Authenticate, and now, you are ready to go!


    ## Deleting Objects
    We do our best to have all our URLs be
    [RESTful](http://en.wikipedia.org/wiki/Representational_state_transfer).
    Every endpoint (URL) may support one of four different http verbs. GET
    requests fetch information about an object, POST requests create objects,
    PUT requests update objects, and finally DELETE requests will delete
    objects.


    ## Structure

    ### The Envelope
    Every response is contained by an envelope. That is, each response has a
...
    ```json
    {
        "meta": {
...
        },
...

...

...
host: api.aswproject.fun
basePath: /
schemes:
...
produces:
...
consumes:
...
securityDefinitions:
  key:
    type: apiKey
    in: header
    name: access_token
security:
  - key: []
  
################################################################################
#                                   Parameters                                 #
################################################################################
parameters:
...
  issue-id:
    name: issue-id
    in: path
    description: An issue identifier number
    type: number
    required: true
...
  priority:
    name: priority
    in: query
    description: Priority of the issue
    type: string
    required: false
    enum:
      - Trivial
      - Minor
      - Major
      - Critical
      - Blocker
  kind:
    name: kind
    in: query
    description: Kind of the issue
    type: string
    required: false
    enum:
      - Bug
      - Enhancement
      - Proposal
      - Task
  status:
    name: status
    in: query
    description: Status of the issue
    type: string
    required: false
    enum:
      - New
      - Open
      - Resolved
  assigned:
    name: assigned
    in: query
    description: Email of the user whom the issue has been assigned to
    type: string
    required: false
  watching:
    name: watching
    in: query
    description: Email of the user watching the issue
    type: string
    required: false
  
################################################################################
#                                     Paths                                    #
################################################################################
paths:
  /users:
    get: 
      security: 
        - key: [] 
      tags: 
        - Users 
      description: |
        Get all users in the system 
      responses: 
        200: 
          description: OK 
          schema: 
            properties:
              meta:
                properties:
                  code:
                    type: integer

              data:
                title: users
                type: array
                items:
                  $ref: '#/definitions/User'

              _links:
                properties:
                  self:
                    type: object
                    properties:
                      href:
                        type: string
        401:
          $ref: '#/responses/Standard401ErrorResponse'
    post:
      tags:
        - Users
      description: |
        Create a new user
      parameters:
        - in: body
          name: User
          description: New user object
          required: false
          schema:
            $ref: "#/definitions/newUser"
      responses:
        201:
          description: Created
          schema:
            properties:
                meta:
                  properties:
                    code:
                      type: integer
                      
                data:
                  title: users
                  type: array
                  items:
                    $ref: '#/definitions/User'
                      
                _links:
                  properties:
                    self:
                      type: object
                      properties:
                        href:
                          type: string
        403:
          $ref: '#/responses/Standard403ErrorResponse'
...
  /issues:
    get:
      security:
        - key: []
      tags:
        - Issues
      description: Get a list with all the issues. They can be filtered using the query parameters.
      parameters:
        - $ref: "#/parameters/priority"
        - $ref: '#/parameters/kind'
        - $ref: '#/parameters/status'
        - $ref: '#/parameters/assigned'
        - $ref: '#/parameters/watching'
      responses:
        200:
          description: OK
          schema:
            properties:
                meta:
                  properties:
                    code:
                      type: integer
                      
                data:
                  title: issues
                  type: array
                  items:
                    $ref: '#/definitions/Issue'
                      
                _links:
                  properties:
                    self:
                      type: object
                      properties:
                        href:
                          type: string
        401:
          $ref: '#/responses/Standard401ErrorResponse'

    post:
      security:
        - key: []
      tags:
        - Issues
      description: Create a new issue
      parameters:
        - in: body
          name: Issue
          description: New issue object
          required: false
          schema:
            $ref: "#/definitions/newIssue"
      responses:
        201:
          description: Created
          schema:
            properties:
                meta:
                  properties:
                    code:
                      type: integer
                      
                data:
                  title: issues
                  type: array
                  items:
                    $ref: '#/definitions/Issue'
                      
                _links:
                  properties:
                    self:
                      type: object
                      properties:
                        href:
                          type: string
   
        401:
          $ref: '#/responses/Standard401ErrorResponse'
        403:
          $ref: '#/responses/Standard403ErrorResponse'
        404:
          $ref: '#/responses/AssigneeDoesNotExistResponse'
...
  /issues/{issue-id}/comments:
    parameters:
    - $ref: '#/parameters/issue-id'
    - in: body
      name: Issue
      description: New issue object
      required: false
      schema:
        $ref: "#/definitions/newComment"
    post:
      security:
        - key: []
      tags:
        - Issues
        - Comments
      description: Create a new comment for an issue
      responses:
        201:
          description: Created
          schema:
            properties:
                meta:
                  properties:
                    code:
                      type: integer
                      
                data:
                  $ref: '#/definitions/Comment'
                      
                _links:
                  properties:
                    self:
                      type: object
                      properties:
                        href:
                          type: string
        401: 
          $ref: '#/responses/Standard401ErrorResponse'
        404: 
          $ref: '#/responses/IssueDoesNotExistResponse'
...
definitions:
  User:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
      email:
        type: string
  Issue:
    type: object
    properties:
      id:
        type: integer
      title:
        type: string
      description:
        type: string
      kind:
        type: string
      priority:
        type: string
      status:
        type: string
      spam:
        type: boolean
      created_at:
        type: string
      updated_at:
        type: string
      comments:
        type: object
        properties:
          count:
            type: integer
          data:
            type: array
            items:
              $ref: '#/definitions/Comment'
      user_creator:
        $ref: '#/definitions/User'
      user_assignee:
        $ref: '#/definitions/User'
  Comment:
    type: object
    properties:
      id:
        type: integer
      text:
        type: string
      created_at:
        type: string
      user_creator:
        $ref: '#/definitions/User'
  newUser:
    type: object
    required:
      - name
      - email
      - password
      - password_confirmation
    properties:
      name:
        type: string
      email:
        type: string
      password:
        type: string
        format: password
      password_confirmation:
        type: string
        format: password
  newIssue:
    type: object
    required:
      - title
      - kind
      - priority
    properties:
      title:
        type: string
        description: Title
      description:
        type: string
        description: Description
      assignee-id:
        type: integer
        description: Id of the user whom the issue will be assigned to
      kind:
        type: string
        enum:
          - Bug
          - Enhancement
          - Proposal
          - Task
      priority:
        type: string
        enum:
          - Trivial
          - Minor
          - Major
          - Critical
          - Blocker
  newComment:
    type: object
    properties:
      text:
        type: string
        description: text
...
responses:
  Standard401ErrorResponse:
    description: Unauthorized
  Standard403ErrorResponse:
    description: Forbidden
...
  IssueDoesNotExistResponse:
    description: An issue with the specified `issue-id` was not found.
...
  AssigneeDoesNotExistResponse:
    description: A user with the specified `assignee-id` was not found.